Disallows you from deleting the currently-logged-in account

+++ src/main/java/edu/wpi/FlashyFrogs/Accounts/LoginAdministratorController.java
@@ -18,6 +18,7 @@ import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Node;
 import javafx.scene.control.Button;
+import javafx.scene.control.Label;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 import javafx.scene.control.cell.PropertyValueFactory;
@@ -27,6 +28,7 @@ import org.hibernate.Session;
 
 @GeneratedExclusion
 public class LoginAdministratorController implements IController {
+  @FXML private Label errorMessage;
   @FXML private TableView<UserLogin> tableView;
   @FXML private TableView<UserLogin> userLoginTable;
   @FXML private TableColumn<UserLogin, Number> idCol;
@@ -72,6 +74,7 @@ public class LoginAdministratorController implements IController {
   }
 
   public void initialize() throws Exception {
+    errorMessage.setVisible(false);
     h1.setVisible(false);
 
     // Clear old table before init
@@ -127,13 +130,18 @@ public class LoginAdministratorController implements IController {
 
   public void deleteUser(ActionEvent actionEvent) {
     Session ses = CONNECTION.getSessionFactory().openSession();
-    ses.beginTransaction();
-    ses.createMutationQuery("delete FROM HospitalUser user WHERE id=:ID")
-        .setParameter("ID", selectedUserLogin.getUser().getId())
-        .executeUpdate();
-    ses.getTransaction().commit();
-    ses.close();
-    userLoginTable.getItems().remove(selectedUserLogin);
+    if (selectedUserLogin.getUser().equals(CurrentUserEntity.CURRENT_USER.getCurrentuser())) {
+      errorMessage.setVisible(true);
+    } else {
+      errorMessage.setVisible(false);
+      ses.beginTransaction();
+      ses.createMutationQuery("delete FROM HospitalUser user WHERE id=:ID")
+          .setParameter("ID", selectedUserLogin.getUser().getId())
+          .executeUpdate();
+      ses.getTransaction().commit();
+      ses.close();
+      userLoginTable.getItems().remove(selectedUserLogin);
+    }
   }
 
   public void setSelected() {
